requirements.txt
python-telegram-bot==20.7
beautifulsoup4==4.12.3
lxml==5.1.0
Pillow==10.2.0
pytest==8.0.0
